requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
schedule
requests
python-dotenv #If you are going to use a .env file